requirements.txt
streamlit
pandas
openpyxl
pyrebase4
pycryptodome
google-cloud-storage
google-cloud-firestore
google-auth
setuptools